feat: guest drawer menu
