+++ catalyst_voices/lib/pages/spaces/drawer/my_private_proposals.dart
@@ -1,32 +1,34 @@
+import 'package:catalyst_voices/pages/spaces/drawer/space_header.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_models/catalyst_voices_models.dart';
 import 'package:flutter/material.dart';
 
-class IndividualPrivateCampaigns extends StatelessWidget {
-  const IndividualPrivateCampaigns({super.key});
+class MyPrivateProposals extends StatelessWidget {
+  const MyPrivateProposals({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Column(
       mainAxisSize: MainAxisSize.min,
       children: [
+        SpaceHeader(Space.workspace),
         SectionHeader(
           leading: SizedBox(width: 12),
-          title: Text('Individual private campaigns'),
+          title: Text('My private proposals (3/5)'),
         ),
         VoicesNavTile(
-          name: 'Fund name 1',
-          status: ProposalStatus.ready,
-          trailing: MoreOptionsButton(),
-        ),
-        VoicesNavTile(
-          name: 'Campaign 1',
+          name: 'My first proposal',
           status: ProposalStatus.draft,
           trailing: MoreOptionsButton(),
         ),
         VoicesNavTile(
-          name: 'What happens with a campaign title that is longer that',
-          status: ProposalStatus.draft,
+          name: 'My second proposal',
+          status: ProposalStatus.inProgress,
+          trailing: MoreOptionsButton(),
+        ),
+        VoicesNavTile(
+          name: 'My third proposal',
+          status: ProposalStatus.inProgress,
           trailing: MoreOptionsButton(),
         ),
       ],
