+++ catalyst_voices/lib/pages/spaces/drawer/voting_rounds.dart
@@ -1,3 +1,4 @@
+import 'package:catalyst_voices/pages/spaces/drawer/space_header.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_assets/catalyst_voices_assets.dart';
 import 'package:catalyst_voices_models/catalyst_voices_models.dart';
@@ -11,6 +12,7 @@ class VotingRounds extends StatelessWidget {
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: [
+        const SpaceHeader(Space.voting),
         const SectionHeader(
           leading: SizedBox(width: 12),
           title: Text('Active funding rounds'),
