+++ catalyst_voices/lib/pages/spaces/spaces_shell_page.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_voices/pages/spaces/spaces_drawer.dart';
+import 'package:catalyst_voices/pages/spaces/drawer/spaces_drawer.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_blocs/catalyst_voices_blocs.dart';
 import 'package:catalyst_voices_models/catalyst_voices_models.dart';
@@ -19,6 +19,7 @@ class SpacesShellPage extends StatelessWidget {
   Widget build(BuildContext context) {
     final sessionBloc = context.watch<SessionBloc>();
     final isVisitor = sessionBloc.state is VisitorSessionState;
+    final isUnlocked = sessionBloc.state is ActiveUserSessionState;
 
     return Scaffold(
       appBar: VoicesAppBar(
@@ -29,7 +30,12 @@ class SpacesShellPage extends StatelessWidget {
           SessionStateHeader(),
         ],
       ),
-      drawer: isVisitor ? null : SpacesDrawer(space: space),
+      drawer: isVisitor
+          ? null
+          : SpacesDrawer(
+              space: space,
+              isUnlocked: isUnlocked,
+            ),
       body: child,
     );
   }
